move wrapper code to its own translation unit
