+++ cpp/box/CMakeLists.txt
@@ -1,5 +1,5 @@
 # create the target and add the needed properties
-nanobind_add_module(_box SHARED module-box.cc export_Box.h)
+nanobind_add_module(_box SHARED module-box.cc export_Box.cc)
 target_link_libraries(_box PUBLIC TBB::tbb)
 # this probably isn't needed for box, but may be needed by other modules
 # target_compile_definitions( box # Avoid deprecation warnings for unsupported
